公卫修改日志 1、	出生医学证明从2014年1月1日更新后，没有家庭住址的输入框，从而导致在出生医学证明查询界面家庭住址出不来，解决办法：在“出生医学证明”表里写了一个“插入、更新”触发器，将父亲的住址同步至家庭住址中。 2、	居民健康档案更新出生日期后，会对业务数据产生影响，如儿童的年龄等，解决办法：重新保存业务数据。 3、	在新生儿访视中，新生儿疾病筛查改为多选。 4、	在新生儿访视中，前卤形状选择“正常”能正常保存，但不能正常显示，解决办法：更新数据库此值的字典表。 5、	在新生儿访视中，脐带选择“未脱”能正常保存，但不能正常显示，解决办法：更新数据库此值的字典表。 6、	在新生儿访视中，父母等信息可以根据出生医学证明信息带出来。 7、	在儿童保健手册中，父母等相关信息可以根据出生医学证明、分娩记录、居民健康档案带出来。 8、	在儿童保健手册中，母亲孕期异常情况和小儿围产期异常情况选择有后不可以再选择其它的，解决办法：更改显示列表的属性。 9、	在第2至5次产检中，界面骨盆测量的数据类型为数值类型，但数据库保存该字段的类型为整型，导致不能正常保存，修改数据库字段类型。 10、	在“出生医学证明使用”中，医疗机构外出生的签发，医疗机构名称改为直接填写，不需要选择。

+++ src/cn/net/tongfang/framework/security/vo/VisitBeforeBorn.java
@@ -44,10 +44,10 @@ public class VisitBeforeBorn implements java.io.Serializable {
 	private String edema;
 	private String diagnosisRemark;
 	private String bornBirthAddressPlan;
-	private Integer pelvis01;
-	private Integer pelvis02;
-	private Integer pelvis03;
-	private Integer pelvis04;
+	private Double pelvis01;
+	private Double pelvis02;
+	private Double pelvis03;
+	private Double pelvis04;
 	private String foreignId;
 	// Constructors
 
@@ -74,8 +74,8 @@ public class VisitBeforeBorn implements java.io.Serializable {
 			String highRiskRemark, String execDistrictNum,
 			String highRiskSearch, Integer gravidity, String edema,
 			String diagnosisRemark, String bornBirthAddressPlan,
-			Integer pelvis01, Integer pelvis02, Integer pelvis03,
-			Integer pelvis04,String foreignId) {
+			Double pelvis01, Double pelvis02, Double pelvis03,
+			Double pelvis04,String foreignId) {
 		super();
 		this.id = id;
 		this.fileNo = fileNo;
@@ -392,35 +392,35 @@ public class VisitBeforeBorn implements java.io.Serializable {
 		this.bornBirthAddressPlan = bornBirthAddressPlan;
 	}
 
-	public Integer getPelvis01() {
+	public Double getPelvis01() {
 		return pelvis01;
 	}
 
-	public void setPelvis01(Integer pelvis01) {
+	public void setPelvis01(Double pelvis01) {
 		this.pelvis01 = pelvis01;
 	}
 
-	public Integer getPelvis02() {
+	public Double getPelvis02() {
 		return pelvis02;
 	}
 
-	public void setPelvis02(Integer pelvis02) {
+	public void setPelvis02(Double pelvis02) {
 		this.pelvis02 = pelvis02;
 	}
 
-	public Integer getPelvis03() {
+	public Double getPelvis03() {
 		return pelvis03;
 	}
 
-	public void setPelvis03(Integer pelvis03) {
+	public void setPelvis03(Double pelvis03) {
 		this.pelvis03 = pelvis03;
 	}
 
-	public Integer getPelvis04() {
+	public Double getPelvis04() {
 		return pelvis04;
 	}
 
-	public void setPelvis04(Integer pelvis04) {
+	public void setPelvis04(Double pelvis04) {
 		this.pelvis04 = pelvis04;
 	}
